# server/internal/api/handlers/openapi.yaml
openapi: 3.0.3
info:
  title: AgentHound API
  description: >-
    REST API for `agenthound-server`. The server binds to 127.0.0.1:8080
    by default and is single-user. Mutating endpoints are gated by a
    localhost token: the server generates a 32-byte token at startup
    and writes it to `~/.agenthound/server.token` (override with
    `AGENTHOUND_TOKEN_PATH`). The embedded React UI fetches the token
    from `GET /api/v1/auth/local-token` on first load and includes it
    in subsequent mutating requests as `Authorization: Bearer <token>`.


    Read-only endpoints (graph reads, findings, prebuilt queries,
    rules, health, docs) are intentionally open. Single-user posture
    means localhost reads are fine; gating them would force the UI to
    plumb auth headers through every read for no security gain.


    The `/ingest` endpoint is consumed primarily by the embedded React
    UI's "Scan Manager → Import scan" drag-drop dialog. The
    `agenthound` collector binary writes JSON to a file or stdout; CLI
    ingest (`agenthound-server ingest <file>`) bypasses HTTP entirely
    and does not require the localhost token.
  version: "0.1.0"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Default

# Mutating endpoints require the localhost Bearer token. Read endpoints
# omit `security` and remain open.
security: []
components:
  securitySchemes:
    LocalhostToken:
      type: http
      scheme: bearer
      bearerFormat: hex
      description: >-
        32-byte hex token loaded from `~/.agenthound/server.token` (or
        `$AGENTHOUND_TOKEN_PATH`). Fetch via `GET /auth/local-token`
        from the same origin; do not embed the token in client code.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: description of the problem
        details:
          description: Additional error context (validation errors, etc.)
...
    Node:
      type: object
      properties:
        id:
          type: string
        kinds:
          type: array
          items:
            type: string
        properties:
          type: object
          additionalProperties: true
    Edge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        kind:
          type: string
        source_kind:
          type: string
        target_kind:
          type: string
        properties:
          type: object
          additionalProperties: true
    NodeDetail:
      type: object
      properties:
        node:
          $ref: "#/components/schemas/Node"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/Edge"
    SearchResultItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
    NeighborhoodResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/Edge"
    BlastRadiusResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/Edge"
        rings:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        direction:
          type: string
          enum: [out, in, both]
        max_hops:
          type: integer
    PathRequest:
      type: object
      required: [source, source_kind]
      properties:
        source:
          type: string
        target:
          type: string
        source_kind:
          type: string
        target_kind:
          type: string
        max_hops:
          type: integer
          default: 10
          minimum: 1
          maximum: 20
        limit:
          type: integer
          default: 10
          minimum: 1
          maximum: 100
    PathNode:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kinds:
          type: array
          items:
            type: string
    PathEdge:
      type: object
      properties:
        kind:
          type: string
        source:
          type: string
        target:
          type: string
        risk_weight:
          type: number
    PathResult:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/PathNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/PathEdge"
        hops:
          type: integer
        weight:
          type: number
    PathResponse:
      type: object
      properties:
        paths:
          type: array
          items:
            $ref: "#/components/schemas/PathResult"
        algorithm:
          type: string
          description: Only present for weighted-path responses
    Finding:
      type: object
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        category:
          type: string
        title:
          type: string
        description:
          type: string
        edge_kind:
          type: string
        source_id:
          type: string
        source_name:
          type: string
        source_kind:
          type: string
        target_id:
          type: string
        target_name:
          type: string
        target_kind:
          type: string
        confidence:
          type: number
        owasp_map:
          type: array
          items:
            type: string
        triage:
          allOf:
            - $ref: "#/components/schemas/TriageState"
          nullable: true
          description: >-
            Inline triage state for this finding (omitted when no decision
            has been recorded). Lets the register render the triage control
            without a per-row round-trip.
    TriageState:
      type: object
      properties:
        status:
          type: string
          enum: [new, triaging, confirmed, accepted-risk, false-positive]
        note:
          type: string
        updated_at:
          type: string
          format: date-time
    TriageUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, triaging, confirmed, accepted-risk, false-positive]
        note:
          type: string
    FindingDetail:
      type: object
      properties:
        finding:
          $ref: "#/components/schemas/Finding"
        evidence:
          type: object
          additionalProperties: true
    PreBuiltQuery:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        severity:
          type: string
        owasp_map:
          type: array
          items:
            type: string
    PreBuiltResult:
      type: object
      properties:
        query:
          $ref: "#/components/schemas/PreBuiltQuery"
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
    Rule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: integer
        enabled:
          type: boolean
        severity:
          type: string
        scope:
          type: object
          properties:
            collector:
              type: string
            targets:
              type: array
              items:
                type: string
        owasp:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: string
    IngestResult:
      type: object
      properties:
        scan_id:
          type: string
        nodes_written:
          type: integer
        edges_written:
          type: integer
        warnings:
          type: array
          items:
            type: string
        post_processing_stats:
          type: array
          items:
            type: object
            properties:
              processor_name:
                type: string
        duration:
          type: integer
          description: Duration in nanoseconds
    Scan:
      type: object
      properties:
        id:
          type: string
        collector:
          type: string
        status:
          type: string
          enum: [pending, running, completed, completed_with_errors, failed]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        node_count:
          type: integer
        edge_count:
          type: integer
        error:
          type: string
        metadata:
          type: object
          additionalProperties: true
    CreateScanRequest:
      type: object
      required: [collector]
      properties:
        collector:
          type: string
          enum: [mcp, a2a, config]
        metadata:
          type: object
          additionalProperties: true
    QueryRequest:
      type: object
      required: [cypher]
      properties:
        cypher:
          type: string
        params:
          type: object
          additionalProperties: true
    QueryResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
  responses:
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Localhost token missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /graph/search:
    get:
      summary: Search graph nodes
      description: Free-text search over node properties.
      operationId: searchGraph
      tags: [Graph]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 10000
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResultItem"
  /graph/nodes:
    get:
      summary: List nodes
      description: List graph nodes with optional kind filter.
      operationId: listNodes
      tags: [Graph]
      parameters:
        - name: kind
          in: query
          schema:
            type: string
          description: Filter by node label (e.g. MCPServer, MCPTool)
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 10000
      responses:
        "200":
          description: Nodes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
  /graph/nodes/{id}:
    get:
      summary: Get node
      description: Returns a single node with all connected edges.
      operationId: getNode
      tags: [Graph]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Node with edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /graph/nodes/{id}/neighborhood:
    get:
      summary: Node neighborhood
      description: Returns the immediate (or N-hop) neighborhood of a node.
      operationId: getNodeNeighborhood
      tags: [Graph]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: depth
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
            maximum: 3
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Neighborhood subgraph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeighborhoodResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /graph/nodes/{id}/blast-radius:
    get:
      summary: Node blast radius
      description: >-
        Returns nodes reachable from (or to) the given node up to
        `max_hops`, grouped by ring distance.
      operationId: getNodeBlastRadius
      tags: [Graph]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
            enum: [out, in, both]
            default: out
        - name: max_hops
          in: query
          schema:
            type: integer
            default: 3
            minimum: 1
            maximum: 10
      responses:
        "200":
          description: Blast radius subgraph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlastRadiusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /graph/edges:
    get:
      summary: List edges
      description: List graph edges with optional filters.
      operationId: listEdges
      tags: [Graph]
      parameters:
        - name: kind
          in: query
          schema:
            type: string
          description: Filter by edge kind (e.g. TRUSTS_SERVER, CAN_REACH)
        - name: source
          in: query
          schema:
            type: string
          description: Filter by source node ID
        - name: target
          in: query
          schema:
            type: string
          description: Filter by target node ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 10000
      responses:
        "200":
          description: Edges list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Edge"
  /ingest:
    post:
      summary: Ingest collector output
      description: >-
        Upload collector JSON. Runs the full pipeline: validate, normalize,
        deduplicate, write to Neo4j, post-process composite edges.
        Max body 100 MB.
      operationId: ingest
      tags: [Ingest]
      security:
        - LocalhostToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Collector JSON output (see graph model docs)
      responses:
        "200":
          description: Ingest successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analysis/shortest-path:
    post:
      summary: Shortest path
      description: Find the shortest path between two nodes in the graph.
      operationId: shortestPath
      tags: [Analysis]
      security:
        - LocalhostToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathRequest"
      responses:
        "200":
          description: Paths found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analysis/all-paths:
    post:
      summary: All paths
      description: Enumerate all paths between two nodes (bounded by max_hops and limit).
      operationId: allPaths
      tags: [Analysis]
      security:
        - LocalhostToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathRequest"
      responses:
        "200":
          description: Paths found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analysis/weighted-path:
    post:
      summary: Weighted (lowest-risk) path
      description: >-
        Find the lowest-risk-weight path using Dijkstra (APOC) or
        shortestPath+reduce fallback. Response includes the algorithm used.
      operationId: weightedPath
      tags: [Analysis]
      security:
        - LocalhostToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathRequest"
      responses:
        "200":
          description: Weighted path found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analysis/findings:
    get:
      summary: Security findings
      description: >-
        List findings from the latest persisted snapshot (one row per
        fingerprint), each with inline triage state. Findings triaged as
        `accepted-risk` or `false-positive` are hidden unless
        `include_suppressed=true`.
      operationId: getFindings
      tags: [Analysis]
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: include_suppressed
          in: query
          description: When true, include accepted-risk / false-positive findings.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Findings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
  /findings/triage/{fingerprint}:
    get:
      summary: Get finding triage state
      description: >-
        Returns the cross-scan triage decision for a finding fingerprint.
        Open read. A fingerprint with no recorded decision returns the
        implicit `new` state.
      operationId: getTriage
      tags: [Analysis]
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
          description: 16-character hex finding fingerprint
      responses:
        "200":
          description: Triage state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriageState"
        "400":
          $ref: "#/components/responses/ValidationError"
    put:
      summary: Set finding triage state
      description: >-
        Records (or updates) the triage decision for a finding fingerprint.
        Triage state has no foreign key to findings, so it survives scan
        deletion and re-detection.
      operationId: setTriage
      tags: [Analysis]
      security:
        - LocalhostToken: []
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
          description: 16-character hex finding fingerprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriageUpdateRequest"
      responses:
        "200":
          description: Updated triage state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriageState"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analysis/findings/{id}:
    get:
      summary: Finding detail
      description: Returns evidence and context for a single finding.
      operationId: getFindingDetail
      tags: [Analysis]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Finding detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /analysis/prebuilt/{id}:
    get:
      summary: Execute pre-built query
      description: Run a pre-built query by ID and return results.
      operationId: runPreBuilt
      tags: [Analysis]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreBuiltResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /scans:
    get:
      summary: List scans
      description: List scan history with pagination.
      operationId: listScans
      tags: [Scans]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Scan list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
    post:
      summary: Create scan
      description: Register a new scan.
      operationId: createScan
      tags: [Scans]
      security:
        - LocalhostToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
      responses:
        "201":
          description: Scan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /scans/{id}:
    get:
      summary: Get scan
      description: Get scan details by ID.
      operationId: getScan
      tags: [Scans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete scan
      description: >-
        Delete a scan from the scan history and remove all graph
        nodes and edges that were exclusively owned by this scan.
      operationId: deleteScan
      tags: [Scans]
      security:
        - LocalhostToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Scan deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /rules:
    get:
      summary: List detection rules
      description: Returns metadata for every loaded detection rule (builtin and custom).
      operationId: listRules
      tags: [Rules]
      parameters:
        - name: collector
          in: query
          schema:
            type: string
            enum: [mcp, a2a, config, all]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Rule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
  /rules/{id}:
    get:
      summary: Get rule
      description: Returns the full definition of a single detection rule.
      operationId: getRule
      tags: [Rules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          $ref: "#/components/responses/NotFound"
  /query:
    post:
      summary: Raw Cypher query
      description: >-
        Execute a raw Cypher query against Neo4j. Gated behind the
        localhost Bearer token; CLI users bypass HTTP entirely via
        `agenthound-server query`.
      operationId: rawQuery
      tags: [Admin]
      security:
        - LocalhostToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"